Add a ton of draw calls, completely unoptimised
Attempting to give this at least a little bit of load and start to identify where we can leverage performance improvements between DX11 and Dx12

We're just drawing everything the same, but across two grids of viewports (the second grid overlaid, but intentionally offset to fill in the gaps)

Each grid is 1000x1000 (two grids totalling 2M viewports), with the two grids offset 50% from each other

We're basically in horrifically poorly designed particle effects territory. Of course, we're not doing any kind of transform here so the vertex and pixel shaders remain obscenely simple

This remains untested on D3D12 (because I'm not on a D3D12 capable machine right this minute)
If this proves to demonstrate any kind of performance degradation, we might have an opportunity to test command recording on multiple threads.

This example does not lend itself to the use of bundles, due to usage of RSSetViewports

Obviously, if we wanted to do this *optimally* we would leverage DrawInstanced and perform layout in-shader using the instanceId

We can crank up the number of viewports easily enough if we need to.
This should demonstrate CPU load at least, if not GPU load